refactor(release): split prepare helpers from pipeline
