+++ src/core/release/pipeline_capabilities.rs
@@ -1,0 +1,24 @@
+use crate::extension::ExtensionManifest;
+
+/// Derive publish targets from extensions that have `release.publish` action.
+pub(super) fn get_publish_targets(extensions: &[ExtensionManifest]) -> Vec<String> {
+    extensions
+        .iter()
+        .filter(|m| m.actions.iter().any(|a| a.id == "release.publish"))
+        .map(|m| m.id.clone())
+        .collect()
+}
+
+/// Check if any extension provides the `release.package` action.
+pub(super) fn has_package_capability(extensions: &[ExtensionManifest]) -> bool {
+    extensions
+        .iter()
+        .any(|m| m.actions.iter().any(|a| a.id == "release.package"))
+}
+
+/// Check if any extension provides the `release.prepare` action.
+pub(super) fn has_prepare_capability(extensions: &[ExtensionManifest]) -> bool {
+    extensions
+        .iter()
+        .any(|m| m.actions.iter().any(|a| a.id == "release.prepare"))
+}
